requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
streamlit>=1.0.0
pydantic>=1.8.2
python-dotenv>=0.19.0
requests>=2.26.0
torch>=1.9.0
transformers>=4.12.0
langgraph>=0.1.0
PyYAML>=5.4.1